+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 #### Added
 
 #### Changed
+ - update H3 library to 4.1.1
 
 #### Removed